+++ cpu/mem/memory.go
@@ -1,18 +1,18 @@
 package mem
 
 type Memory interface {
-	ReadByteAt(virtualAddress uint64) byte
-	ReadHalfWordAt(virtualAddress uint64) uint16
-	ReadWordAt(virtualAddress uint64) uint32
-	ReadDoubleWordAt(virtualAddress uint64) uint64
-	ReadBlockAt(virtualAddress uint64, size uint64) []byte
-	ReadStringAt(virtualAddress uint64, size uint64) string
-	WriteByteAt(virtualAddress uint64, data byte)
-	WriteHalfWordAt(virtualAddress uint64, data uint16)
-	WriteWordAt(virtualAddress uint64, data uint32)
-	WriteDoubleWordAt(virtualAddress uint64, data uint64)
-	WriteStringAt(virtualAddress uint64, data string)
-	WriteBlockAt(virtualAddress uint64, size uint64, data []byte)
+	ReadByteAt(virtualAddress uint32) byte
+	ReadHalfWordAt(virtualAddress uint32) uint16
+	ReadWordAt(virtualAddress uint32) uint32
+	ReadDoubleWordAt(virtualAddress uint32) uint64
+	ReadBlockAt(virtualAddress uint32, size uint32) []byte
+	ReadStringAt(virtualAddress uint32, size uint32) string
+	WriteByteAt(virtualAddress uint32, data byte)
+	WriteHalfWordAt(virtualAddress uint32, data uint16)
+	WriteWordAt(virtualAddress uint32, data uint32)
+	WriteDoubleWordAt(virtualAddress uint32, data uint64)
+	WriteStringAt(virtualAddress uint32, data string)
+	WriteBlockAt(virtualAddress uint32, size uint32, data []byte)
 }
 
 type MemoryReader interface {
@@ -20,8 +20,8 @@ type MemoryReader interface {
 	ReadHalfWord() uint16
 	ReadWord() uint32
 	ReadDoubleWord() uint64
-	ReadString(size uint64) string
-	ReadBlock(size uint64) []byte
+	ReadString(size uint32) string
+	ReadBlock(size uint32) []byte
 }
 
 type MemoryWriter interface {
@@ -30,5 +30,5 @@ type MemoryWriter interface {
 	WriteWord(data uint32)
 	WriteDoubleWord(data uint64)
 	WriteString(data string)
-	WriteBlock(size uint64, data []byte)
+	WriteBlock(size uint32, data []byte)
 }
